+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Atlas Android
 
+## 0.4.11
+
+### Major Changes
+  * Update to Layer Android SDK Version 0.23.17
+  * Added read/unread styling for DateTextView
+  * Added getter/setter for avatar presence to the ConversationsRecyclerView
+
 ## 0.4.10
 
 ### Major Changes